+++ imio/dms/mail/adapters.py
@@ -1177,6 +1177,43 @@ class ItemSignersAdapter(object):
         #         yield sub_content.UID()
 
 
+class BaseItemOrderProvider(object):
+    """Group children by portal_type: ged files most recent first, appendices by title."""
+
+    portal_types = ()
+    alphabetical_types = ("dmsappendixfile",)
+
+    def __init__(self, context):
+        self.context = context
+
+    def get_item_order(self):
+        order = {}
+        idx = 0
+        for portal_type in self.portal_types:
+            children = [
+                obj for obj in self.context.objectValues()
+                if getattr(obj, "portal_type", None) == portal_type
+            ]
+            if portal_type in self.alphabetical_types:
+                children = sorted(children, key=lambda obj: safe_unicode(obj.Title()).lower())
+            else:
+                children = reversed(children)
+            for obj in children:
+                order[obj.UID()] = idx
+                idx += 1
+        return order
+
+
+class IMItemOrderProvider(BaseItemOrderProvider):
+
+    portal_types = ("dmsmainfile", "dmsappendixfile")
+
+
+class OMItemOrderProvider(BaseItemOrderProvider):
+
+    portal_types = ("dmsommainfile", "dmsappendixfile")
+
+
 @implementer(ILocalRoleProvider)
 class ApprovalRoleAdapter(object):
     """borg.localrole adapter to set localrole for signing approvers and signers"""